web/requirements.txt
adal==1.2.3
astroid==2.4.1
azure-common==1.1.25
azure-servicebus==7.12.2
certifi==2020.4.5.1
cffi==1.16.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography>=3.4.8
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
idna==2.9
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.13
msrestazure==0.6.3
oauthlib==3.1.0
psycopg2-binary==2.9.9
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-http-client==3.2.7
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.3.0
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
uamqp==1.6.9
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
